requirements.txt
litellm
openai
fastapi
uvicorn
bs4
requests
<<<<<<< HEAD
redis
uuid
streamlit
=======
docling
pandas
numpy
boto3
datetime
pydantic
pathlib
pymupdf
azure-ai-documentintelligence
scrapy
twisted
>>>>>>> origin/main
